Write geo_public when writing other meta fields.
WP.com themes looks for the presence of geo_public and check that it is
truthy before showing geo-tagging data on the frontend.  So, we can
never just write coordinates, if we want the post to display this
information publicy.  Instead, we now always include geo_public in our
meta writes.

We do this at the time that coordinates and/or the geo_address field are
being written to avoid doing a write of geo_public when the user hasn't
even set a location yet.