Ensure PostMetadata.geoLabel() returns string.
Previously, geoLabel() would return undefined when no meta field was set
for the supplied post.  This interacted poorly with the Search
component, causing it to remain populated even after we cleared the meta
field.  Search distinguishes between undefined and empty string because
it wants to detect when you've supplied a value prop versus when you've
left it off altogether.

Now that we match the Search component's expectations, we no longer need
to manually clear the input.  Instead that value prop will automatically
have the desired effect.  This allows us to eliminate refs in
EditorLocation and EditorLocationSearch, simplifying those components
quite a bit.