Omit initial assignment of geo_address label.
When setting the location via the browser's geolocation API, we were
previously setting geo_address to the returned lat/long coordinates
immediately, before we tried to get a better label by reverse geo-coding
with Google Maps.  This sometimes caused a brief flicker in the UI where
you'd see the coordinates in the search input until Google returned its
results.

Now, we'll continue showing the "Locating..." text until Google returns.
If Google fails, we still fall back to using the coordinates as a label.