Invert the control when Cart Store subscribes to selected site Redux updates
Instead of a global middleware in the `build` chunk importing the Cart Store and
calling its methods, let the Cart Store itself subscribe to Redux store when it's
loaded. Avoids bundling a Cart Store dependency in the `build` chunk and defers
the subscription until the moment when the Cart Store gets initialized.