Data: Extract `httpData` into new `packages/js-requests` library
The `httpData` library was developed inside Calypso serving Calypso's
needs. After some initial API work and testing it's clear that it needs
to live independently as its own library.

This patch creates a new library inside of Calypso's repository and will
serve as the home for any iteration on the design of `httpData`, renamed
as `js-requests`